refactor of {extend,normal}/softlabels
